+++ include/ensmallen_bits/ada_belief/ada_belief_update.hpp
@@ -1,55 +1,52 @@
 /**
- * @file adam_update.hpp
- * @author Ryan Curtin
- * @author Vasanth Kalingeri
+ * @file ada_belief_update.hpp
  * @author Marcus Edel
- * @author Vivek Pal
  *
- * Adam optimizer. Adam is an an algorithm for first-order gradient-based
- * optimization of stochastic objective functions, based on adaptive estimates
- * of lower-order moments.
+ * AdaBelief optimizer update policy. The intuition for AdaBelief is to adapt
+ * the stepsize according to the "belief" in the current gradient direction.
  *
  * ensmallen is free software; you may redistribute it and/or modify it under
  * the terms of the 3-clause BSD license.  You should have received a copy of
  * the 3-clause BSD license along with ensmallen.  If not, see
  * http://www.opensource.org/licenses/BSD-3-Clause for more information.
  */
-#ifndef ENSMALLEN_ADAM_ADAM_UPDATE_HPP
-#define ENSMALLEN_ADAM_ADAM_UPDATE_HPP
+#ifndef ENSMALLEN_ADA_BELIEF_ADA_BELIEF_UPDATE_HPP
+#define ENSMALLEN_ADA_BELIEF_ADA_BELIEF_UPDATE_HPP
 
 namespace ens {
 
 /**
- * Adam is an optimizer that computes individual adaptive learning rates for
- * different parameters from estimates of first and second moments of the
- * gradients as given in the section 7 of the following paper.
+ * The intuition for AdaBelief is to adapt the stepsize according to the
+ * "belief" in the current gradient direction.
  *
  * For more information, see the following.
  *
  * @code
- * @article{Kingma2014,
- *   author  = {Diederik P. Kingma and Jimmy Ba},
- *   title   = {Adam: {A} Method for Stochastic Optimization},
- *   journal = {CoRR},
- *   year    = {2014},
- *   url     = {http://arxiv.org/abs/1412.6980}
+ * @misc{zhuang2020adabelief,
+ *   title         = {AdaBelief Optimizer: Adapting Stepsizes by the Belief in
+ *                    Observed Gradients},
+ *   author        = {Juntang Zhuang and Tommy Tang and Sekhar Tatikonda and
+ *                    Nicha Dvornek and Yifan Ding and Xenophon Papademetris
+ *                    and James S. Duncan},
+ *   year          = {2020},
+ *   eprint        = {2010.07468},
+ *   archivePrefix = {arXiv},
  * }
  * @endcode
  */
-class AdamUpdate
+class AdaBeliefUpdate
 {
  public:
   /**
-   * Construct the Adam update policy with the given parameters.
+   * Construct the AdaBelief update policy with the given parameters.
    *
-   * @param epsilon The epsilon value used to initialise the squared gradient
-   *        parameter.
-   * @param beta1 The smoothing parameter.
-   * @param beta2 The second moment coefficient.
+   * @param epsilon A small constant for numerical stability.
+   * @param beta1 The exponential decay rate for the 1st moment estimates.
+   * @param beta2 The exponential decay rate for the 2nd moment estimates.
    */
-  AdamUpdate(const double epsilon = 1e-8,
-             const double beta1 = 0.9,
-             const double beta2 = 0.999) :
+  AdaBeliefUpdate(const double epsilon = 1e-8,
+                  const double beta1 = 0.9,
+                  const double beta2 = 0.999) :
     epsilon(epsilon),
     beta1(beta1),
     beta2(beta2)
@@ -57,19 +54,19 @@ class AdamUpdate
     // Nothing to do.
   }
 
-  //! Get the value used to initialise the squared gradient parameter.
+  //! Get the value for numerical stability.
   double Epsilon() const { return epsilon; }
-  //! Modify the value used to initialise the squared gradient parameter.
+  //! Modify the value used for numerical stability.
   double& Epsilon() { return epsilon; }
 
-  //! Get the smoothing parameter.
+  //! Get the exponential decay rate for the 1st moment estimates.
   double Beta1() const { return beta1; }
-  //! Modify the smoothing parameter.
+  //! Modify the exponential decay rate for the 1st moment estimates.
   double& Beta1() { return beta1; }
 
-  //! Get the second moment coefficient.
+  //! Get the exponential decay rate for the 2nd moment estimates.
   double Beta2() const { return beta2; }
-  //! Modify the second moment coefficient.
+  //! Modify the exponential decay rate for the 2nd moment estimates.
   double& Beta2() { return beta2; }
 
   /**
@@ -86,20 +83,20 @@ class AdamUpdate
      * This constructor is called by the SGD Optimize() method before the start
      * of the iteration update process.
      *
-     * @param parent AdamUpdate object.
+     * @param parent AdaBeliefUpdate object.
      * @param rows Number of rows in the gradient matrix.
      * @param cols Number of columns in the gradient matrix.
      */
-    Policy(AdamUpdate& parent, const size_t rows, const size_t cols) :
+    Policy(AdaBeliefUpdate& parent, const size_t rows, const size_t cols) :
         parent(parent),
         iteration(0)
     {
       m.zeros(rows, cols);
-      v.zeros(rows, cols);
+      s.zeros(rows, cols);
     }
 
     /**
-     * Update step for Adam.
+     * Update step for AdaBelief.
      *
      * @param iterate Parameters that minimize the function.
      * @param stepSize Step size to be used for the given iteration.
@@ -112,34 +109,29 @@ class AdamUpdate
       // Increment the iteration counter variable.
       ++iteration;
 
-      // And update the iterate.
       m *= parent.beta1;
       m += (1 - parent.beta1) * gradient;
 
-      v *= parent.beta2;
-      v += (1 - parent.beta2) * (gradient % gradient);
+      s *= parent.beta2;
+      s += (1 - parent.beta2) * arma::pow(gradient - m, 2.0) + parent.epsilon;
 
       const double biasCorrection1 = 1.0 - std::pow(parent.beta1, iteration);
       const double biasCorrection2 = 1.0 - std::pow(parent.beta2, iteration);
 
-      /**
-       * It should be noted that the term, m / (arma::sqrt(v) + eps), in the
-       * following expression is an approximation of the following actual term;
-       * m / (arma::sqrt(v) + (arma::sqrt(biasCorrection2) * eps).
-       */
-      iterate -= (stepSize * std::sqrt(biasCorrection2) / biasCorrection1) *
-          m / (arma::sqrt(v) + parent.epsilon);
+      // And update the iterate.
+      iterate -= ((m / biasCorrection1) * stepSize) / (arma::sqrt(s /
+          biasCorrection2) + parent.epsilon);
     }
 
    private:
-    // Instantiated parent object.
-    AdamUpdate& parent;
+    //! Instantiated parent object.
+    AdaBeliefUpdate& parent;
 
-    // The exponential moving average of gradient values.
+    //! The exponential moving average of gradient values.
     GradType m;
 
     // The exponential moving average of squared gradient values.
-    GradType v;
+    GradType s;
 
     // The number of iterations.
     size_t iteration;
@@ -149,10 +141,10 @@ class AdamUpdate
   // The epsilon value used to initialise the squared gradient parameter.
   double epsilon;
 
-  // The smoothing parameter.
+  // The xponential decay rate for the 1st moment estimates.
   double beta1;
 
-  // The second moment coefficient.
+  // The exponential decay rate for the 2nd moment estimates.
   double beta2;
 };